added 2 files for bag of words
added:
USConst_bagOfWords.py: python program using CountVectorizer, etc., to create a bag of words from the US Constitution.  A dictionary of words and their frequencies are printed to an output file along in alphabetical order along with the total number of significant tokens (words) and the top 50 most frequently used words.  
USConst_wordFreq.txt: a text document containing the bag of words from the constitution order alphabetically with their counts, a total count of significant tokens (words) in the US Constitution and the top 50 most frequently used words.

NEXT: 
* Trim down the occurrence of similar words by performing some stemming, etc., to only get root words instead of counting multiple versions of the same word, such as state, states, etc.  
* N-grams in order to catch important phrases, titles, etc., such as Vice President as distinguished from President.  And also to analyze common phrases and to start on topic identification.  
* Topic modeling
* BUT WAIT, THERE'S MORE: Similar task++ with gensim, ntlk, etc., etc.